Move Events feature to vertical-slice folders
Pilot continues with the Events slice (2 endpoints — POST /api/events
to create an event, GET /api/events/{id}/videos to list its videos).
The new Features/Events/EventsController owns just those two
endpoints; the existing Controllers/EventsController.cs keeps the
four access-routed endpoints (GetAllEventsAndGroups, GetAssigned,
RequestAccess, manage requests) and will be renamed when the Access
slice migrates.

- IEventService moved out of Domain.Services into
  Application.Features.Events.
- EventRoutes extracted from ApiEndpoints.cs.
- DI registration moved into per-feature AddEventsFeature()
  extension in EventsModule.cs.
- CreateNewEventRequest moved to Contracts/Features/Events/;
  ContractMappers and the leftover access controller gain the
  matching usings.
- Mobile DanceHttpApiClient gains a using for the new
  Contracts.Features.Events namespace.

EventsAndGroupsResponse, UserEventsAndGroupsResponse and the access
DTOs stay in Contracts/Responses + Contracts/Requests for the Access
slice to migrate.
