+++ src/backend/Application/Features/Events/EventService.cs
@@ -1,8 +1,7 @@
 ﻿using Domain.Entities;
-using Domain.Services;
 using Microsoft.EntityFrameworkCore;
 
-namespace Application.Services;
+namespace Application.Features.Events;
 public class EventService : IEventService
 {
     private readonly IApplicationContext dbContext;
