+++ src/backend/Application/Features/Events/EventsModule.cs
@@ -1,0 +1,15 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Application.Features.Events;
+
+public static class EventsModule
+{
+    extension(IServiceCollection services)
+    {
+        public IServiceCollection AddEventsFeature()
+        {
+            services.AddScoped<IEventService, EventService>();
+            return services;
+        }
+    }
+}
