+++ src/tests/TB.DanceDance.Tests/Features/Events/EventServiceTests.cs
@@ -1,9 +1,9 @@
-﻿using Application.Services;
+﻿using Application.Features.Events;
 using Domain.Entities;
 using Infrastructure.Data;
 using TB.DanceDance.Tests.TestsFixture;
 
-namespace TB.DanceDance.Tests.Application;
+namespace TB.DanceDance.Tests.Features.Events;
 
 public class EventServiceTests : BaseTestClass
 {
